+++ assets/voxygen/shaders/include/globals.glsl
@@ -12,4 +12,5 @@ uniform u_globals {
 	uvec4 light_shadow_count;
 	uvec4 medium;
 	ivec4 select_pos;
+	vec4 gamma;
 };